+++ desktop/js/naturalLight.js
@@ -86,18 +86,18 @@ function addCmdToTable(_cmd) {
   })
 }
 
-$(".eqLogic").delegate(".listCmdInfo", 'click', function () {
-  var el = $(this).closest('.form-group').find('.eqLogicAttr');
-  jeedom.cmd.getSelectModal({ cmd: { type: 'info', subType: 'binary' } }, function (result) {
-    if (el.attr('data-concat') == 1) {
-      el.atCaret('insert', result.human);
-    } else {
-      el.value(result.human);
-    }
-  });
-});
+// $(".eqLogic").delegate(".listCmdInfoBinary", 'click', function () {
+//   var el = $(this).closest('.form-group').find('.eqLogicAttr');
+//   jeedom.cmd.getSelectModal({ cmd: { type: 'info', subType: 'binary' } }, function (result) {
+//     if (el.attr('data-concat') == 1) {
+//       el.atCaret('insert', result.human);
+//     } else {
+//       el.value(result.human);
+//     }
+//   });
+// });
 
-$(".eqLogic").delegate(".listCmdInfoCondition", 'click', function () {
+$(".eqLogic").delegate(".listCmdInfo", 'click', function () {
   var el = $(this).closest('.form-group').find('.eqLogicAttr');
   jeedom.cmd.getSelectModal({ cmd: { type: 'info' } }, function (result) {
     if (el.attr('data-concat') == 1) {